Python/requirements.txt
Appium_Python_Client==2.9.0
selenium==4.8.3
